production  Build at 11-26-2016 at 20.34.11

+++ 3.chunk.js
@@ -1272,7 +1272,7 @@ var FormulaService = (function () {
             if (rawFormula[i] == 'Q') {
                 i++;
                 currentQuesNumber = '';
-                while (!isNaN(parseInt(rawFormula[i])))
+                while (!isNaN(parseFloat(rawFormula[i])))
                     currentQuesNumber += rawFormula[i++];
                 var currentQuesObject = this.jsonBuilderHelper.getTemplateQuestionareWithEmittedLeadFormQuestion()[currentQuesNumber - 1];
                 //for marking as mandatory
@@ -1301,7 +1301,7 @@ var FormulaService = (function () {
             if (rawFormula[i] == 'Q') {
                 i++;
                 currentQuesNumber = '';
-                while (!isNaN(parseInt(rawFormula[i])))
+                while (!isNaN(parseFloat(rawFormula[i])))
                     currentQuesNumber += rawFormula[i++];
                 var currentQuesObject = this.jsonBuilderHelper.getTemplateQuestionareWithEmittedLeadFormQuestion()[currentQuesNumber - 1];
                 if (currentQuesObject) {
@@ -1330,7 +1330,7 @@ var FormulaService = (function () {
             this.updateFormulaValidity(this.jsonBuilderHelper.getJSONBuilt().formula[formula].result, formula);
             if (!this.jsonBuilderHelper.getJSONBuilt().formula[formula].isValid) {
                 areAllFormulasValid = false;
-                allInvalidFormulas += 'Result ' + (parseInt(formula) + 1) + ',';
+                allInvalidFormulas += 'Result ' + (parseFloat(formula) + 1) + ',';
             }
         }
         if (areAllFormulasValid)
@@ -1349,12 +1349,12 @@ var FormulaService = (function () {
                 if (rawFormula[i] == 'Q') {
                     i++;
                     var currentQuesNumber = '';
-                    while (!isNaN(parseInt(rawFormula[i])))
+                    while (!isNaN(parseFloat(rawFormula[i])))
                         currentQuesNumber += rawFormula[i++];
                     var currentQuesObject = this.jsonBuilderHelper.getTemplateQuestionareWithEmittedLeadFormQuestion()[currentQuesNumber - 1];
                     if (!currentQuesObject || ((currentQuesObject.type == 'textfield' && (currentQuesObject.config.type == 'text' || currentQuesObject.config.type == 'email'))
                         || currentQuesObject.type == 'text-area')) {
-                        errorResultList += 'Result ' + (parseInt(formula) + 1) + ',';
+                        errorResultList += 'Result ' + (parseFloat(formula) + 1) + ',';
                         break;
                     }
                 }
@@ -1372,12 +1372,12 @@ var FormulaService = (function () {
                 if (rawFormula[i] == 'Q') {
                     i++;
                     var currentQuesNumber = '';
-                    while (!isNaN(parseInt(rawFormula[i])))
+                    while (!isNaN(parseFloat(rawFormula[i])))
                         currentQuesNumber += rawFormula[i++];
                     var currentQuesObject = this.jsonBuilderHelper.getTemplateQuestionareWithEmittedLeadFormQuestion()[currentQuesNumber - 1];
                     if ((currentQuesObject.type == 'textfield' && (currentQuesObject.config.type == 'text' || currentQuesObject.config.type == 'email'))
                         || currentQuesObject.type == 'text-area') {
-                        errorResultList += 'Result ' + (parseInt(formula) + 1) + ',';
+                        errorResultList += 'Result ' + (parseFloat(formula) + 1) + ',';
                         break;
                     }
                 }
@@ -1401,7 +1401,7 @@ var FormulaService = (function () {
             finalAnswer = 0;
         }
         if (finalAnswer == undefined)
-            finalAnswer = '{R' + (parseInt(formulaIndex) + 1) + '}';
+            finalAnswer = '{R' + (parseFloat(formulaIndex) + 1) + '}';
         var quesNowObject = this.jsonBuilderHelper.getJSONBuilt().formula[formulaIndex];
         var lower, upper;
         if (quesNowObject.range.status) {
@@ -1484,7 +1484,7 @@ var FormulaService = (function () {
             if (genericQuestion[i] == 'Q') {
                 j = ++i;
                 currentQuesNumber = '';
-                while (!isNaN(parseInt(genericQuestion[i])))
+                while (!isNaN(parseFloat(genericQuestion[i])))
                     currentQuesNumber += genericQuestion[i++];
                 genericQuestion = genericQuestion.substring(0, j - 1) +
                     this.getValueOfQuestionNumber(currentQuesNumber) +
