+++ 4.chunk.js
@@ -1501,7 +1501,7 @@ var FormulaService = (function () {
             if (rawFormula[i] == 'Q') {
                 i++;
                 currentQuesNumber = '';
-                while (!isNaN(parseInt(rawFormula[i])))
+                while (!isNaN(parseFloat(rawFormula[i])))
                     currentQuesNumber += rawFormula[i++];
                 var currentQuesObject = this.jsonBuilderHelper.getTemplateQuestionareWithEmittedLeadFormQuestion()[currentQuesNumber - 1];
                 //for marking as mandatory
@@ -1530,7 +1530,7 @@ var FormulaService = (function () {
             if (rawFormula[i] == 'Q') {
                 i++;
                 currentQuesNumber = '';
-                while (!isNaN(parseInt(rawFormula[i])))
+                while (!isNaN(parseFloat(rawFormula[i])))
                     currentQuesNumber += rawFormula[i++];
                 var currentQuesObject = this.jsonBuilderHelper.getTemplateQuestionareWithEmittedLeadFormQuestion()[currentQuesNumber - 1];
                 if (currentQuesObject) {
@@ -1559,7 +1559,7 @@ var FormulaService = (function () {
             this.updateFormulaValidity(this.jsonBuilderHelper.getJSONBuilt().formula[formula].result, formula);
             if (!this.jsonBuilderHelper.getJSONBuilt().formula[formula].isValid) {
                 areAllFormulasValid = false;
-                allInvalidFormulas += 'Result ' + (parseInt(formula) + 1) + ',';
+                allInvalidFormulas += 'Result ' + (parseFloat(formula) + 1) + ',';
             }
         }
         if (areAllFormulasValid)
@@ -1578,12 +1578,12 @@ var FormulaService = (function () {
                 if (rawFormula[i] == 'Q') {
                     i++;
                     var currentQuesNumber = '';
-                    while (!isNaN(parseInt(rawFormula[i])))
+                    while (!isNaN(parseFloat(rawFormula[i])))
                         currentQuesNumber += rawFormula[i++];
                     var currentQuesObject = this.jsonBuilderHelper.getTemplateQuestionareWithEmittedLeadFormQuestion()[currentQuesNumber - 1];
                     if (!currentQuesObject || ((currentQuesObject.type == 'textfield' && (currentQuesObject.config.type == 'text' || currentQuesObject.config.type == 'email'))
                         || currentQuesObject.type == 'text-area')) {
-                        errorResultList += 'Result ' + (parseInt(formula) + 1) + ',';
+                        errorResultList += 'Result ' + (parseFloat(formula) + 1) + ',';
                         break;
                     }
                 }
@@ -1601,12 +1601,12 @@ var FormulaService = (function () {
                 if (rawFormula[i] == 'Q') {
                     i++;
                     var currentQuesNumber = '';
-                    while (!isNaN(parseInt(rawFormula[i])))
+                    while (!isNaN(parseFloat(rawFormula[i])))
                         currentQuesNumber += rawFormula[i++];
                     var currentQuesObject = this.jsonBuilderHelper.getTemplateQuestionareWithEmittedLeadFormQuestion()[currentQuesNumber - 1];
                     if ((currentQuesObject.type == 'textfield' && (currentQuesObject.config.type == 'text' || currentQuesObject.config.type == 'email'))
                         || currentQuesObject.type == 'text-area') {
-                        errorResultList += 'Result ' + (parseInt(formula) + 1) + ',';
+                        errorResultList += 'Result ' + (parseFloat(formula) + 1) + ',';
                         break;
                     }
                 }
@@ -1630,7 +1630,7 @@ var FormulaService = (function () {
             finalAnswer = 0;
         }
         if (finalAnswer == undefined)
-            finalAnswer = '{R' + (parseInt(formulaIndex) + 1) + '}';
+            finalAnswer = '{R' + (parseFloat(formulaIndex) + 1) + '}';
         var quesNowObject = this.jsonBuilderHelper.getJSONBuilt().formula[formulaIndex];
         var lower, upper;
         if (quesNowObject.range.status) {
@@ -1713,7 +1713,7 @@ var FormulaService = (function () {
             if (genericQuestion[i] == 'Q') {
                 j = ++i;
                 currentQuesNumber = '';
-                while (!isNaN(parseInt(genericQuestion[i])))
+                while (!isNaN(parseFloat(genericQuestion[i])))
                     currentQuesNumber += genericQuestion[i++];
                 genericQuestion = genericQuestion.substring(0, j - 1) +
                     this.getValueOfQuestionNumber(currentQuesNumber) +
